requirements.txt
fastapi
uvicorn
transformers
torch
pandas
pymongo
python-dotenv
spacy